Add support for multiple ICO token withdrawals

+++ Sources/CoinTrackingExporterKit/Common/UInt64+Hexadecimal.swift
@@ -1,11 +1,11 @@
 import Foundation
 
-extension Int {
+extension UInt64 {
     
-    static func make(hexadecimal string: String) throws -> Int {
+    static func make(hexadecimal string: String) throws -> UInt64 {
         let string = string.hasPrefix("0x") ? String(string.dropFirst(2)) : string
         
-        guard let int = Int(string, radix: 16) else {
+        guard let int = UInt64(string, radix: 16) else {
             throw "Invalid hexadecimal \(string)"
         }
